scripts/Rasberry_pi-Linux/requirements.txt
pyserial==3.5
pygame==2.1.3